backend/requirements.txt
flask
flask-Cors
gunicorn
beautifulsoup4
google-genai
python-dotenv
extract-msg
